review move fill_missing_ids
