+++ cg/services/orders/store_order_services/store_metagenome_order.py
@@ -123,13 +123,3 @@ class StoreMetagenomeOrderService(StoreOrderService):
         self.status_db.session.add_all(new_samples)
         self.status_db.session.commit()
         return new_samples
-
-    @staticmethod
-    def _fill_in_sample_ids(samples: list[dict], lims_map: dict, id_key: str = "internal_id"):
-        """Fill in LIMS sample ids."""
-        for sample in samples:
-            LOG.debug(f"{sample['name']}: link sample to LIMS")
-            if not sample.get(id_key):
-                internal_id = lims_map[sample["name"]]
-                LOG.info(f"{sample['name']} -> {internal_id}: connect sample to LIMS")
-                sample[id_key] = internal_id
